Renames “env” to plugins 📝
