+++ src/plugins/karma/index.js
