+++ src/plugins/path/index.js
