+++ src/plugins/webpack/babel-plugin.js
