+++ src/plugins/webpack/css-modules-plugin.js
