+++ src/plugins/webpack/index.spec.js
