+++ src/plugins/webpack/media-plugin.js
