+++ src/plugins/webpack/webpack-base-plugin.js
